+++ .github/workflows/python-app.yml
@@ -5,7 +5,6 @@ name: Run pytest
 
 on:
   push:
-  pull_request:
 
 jobs:
   build:
