+++ hvor/utils.py
@@ -1,9 +1,7 @@
-import bz2
-from functools import lru_cache
-from pathlib import Path
-
 import geopandas as gpd
 import pandas as pd
+
+from hvor.data import _load_fylkesdata, _load_kommunedata
 
 
 def _get_kommune_metadata(coordinates, lat_key, lon_key) -> pd.DataFrame:
@@ -37,27 +35,3 @@ def _get_fylke_metadata(coordinates, lat_key, lon_key) -> pd.DataFrame:
         .drop(["geometry", "index_right"], axis=1)
         .to_dict(orient="list")
     )
-
-
-@lru_cache(maxsize=None)
-def _load_kommunedata():
-    """Loads data for kommuner and kommunenummer"""
-    filepath = (
-        Path(__file__).joinpath("../resources/Kommuner-2020-large.json.bz2").resolve()
-    )
-    with bz2.open(filepath, mode="rt", encoding="utf-8") as f:
-        return gpd.read_file(f).rename(
-            mapper={"kommunenummer": "kommunenummer", "navn": "kommune"}, axis=1
-        )
-
-
-@lru_cache(maxsize=None)
-def _load_fylkesdata():
-    """Loads data for fylker and fylkesnummer"""
-    filepath = (
-        Path(__file__).joinpath("../resources/Fylker-2020-large.json.bz2").resolve()
-    )
-    with bz2.open(filepath, mode="rt", encoding="utf-8") as f:
-        return gpd.read_file(f).rename(
-            mapper={"fylkesnummer": "fylkesnummer", "navn": "fylke"}, axis=1
-        )
